classwork

+++ Exceptions/ConverterProgram.cs
@@ -1,10 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.Globalization;
 using System.IO;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using System.Xml;
 using System.Xml.Serialization;
 using NLog;
 
@@ -19,7 +21,8 @@ namespace Exceptions
 			try
 			{
 				var filenames = args.Any() ? args : new[] { "text.txt" };
-				var settings = LoadSettings();
+				Settings settings;
+				settings = LoadSettings();
 				ConvertFiles(filenames, settings);
 			}
 			catch (Exception e)
@@ -31,40 +34,65 @@ namespace Exceptions
 		private static void ConvertFiles(string[] filenames, Settings settings)
 		{
 			var tasks = filenames
-				.Select(fn => Task.Run(() => ConvertFile(fn, settings))) 
+				.Select(fn => Task.Run(() => ConvertFile(fn, settings)).ContinueWith(t => { HandleExceptions(t); })) 
 				.ToArray();
 			Task.WaitAll(tasks); 
 		}
 
+		private static void HandleExceptions(Task task)
+		{
+			if (task.Exception == null)
+				return;
+			foreach (var e in task.Exception.InnerExceptions)
+				log.Error(e);
+		}
+
 		private static Settings LoadSettings() 
 		{
-			var serializer = new XmlSerializer(typeof(Settings));
-			var content = File.ReadAllText("settings.xml");
-			return (Settings) serializer.Deserialize(new StringReader(content));
+			try
+			{
+				var serializer = new XmlSerializer(typeof(Settings));
+				if (!File.Exists("settings.xml"))
+				{
+					log.Warn("Файл настроек settings.xml отсутствует");
+					return Settings.Default;
+				}
+				var content = File.ReadAllText("settings.xml");
+				return (Settings) serializer.Deserialize(new StringReader(content));
+			}
+			catch (Exception e)
+			{
+				throw new XmlException("Не удалось прочитать файл настроек");
+			}
 		}
 
 		private static void ConvertFile(string filename, Settings settings)
 		{
+			if (!File.Exists(filename))
+			{
+				log.Error(new FileNotFoundException($"Не удалось сконвертировать {filename}"));
+				return;
+			}
 			Thread.CurrentThread.CurrentCulture = new CultureInfo(settings.SourceCultureName);
 			if (settings.Verbose)
 			{
 				log.Info("Processing file " + filename);
 				log.Info("Source Culture " + Thread.CurrentThread.CurrentCulture.Name);
 			}
-			IEnumerable<string> lines;
-			try 
+			var lines = PrepareLines(filename); 
+			try
 			{
-				lines = PrepareLines(filename); 
+				var convertedLines = lines
+					.Select(ConvertLine)
+					.Select(s => s.Length + " " + s);
+				File.WriteAllLines(filename + ".out", convertedLines);
 			}
 			catch
 			{
-				log.Error($"File {filename} not found"); 
-				return;
+				throw new FormatException("Некорректная строка");
 			}
-			var convertedLines = lines
-				.Select(ConvertLine)
-				.Select(s => s.Length + " " + s);
-			File.WriteAllLines(filename + ".out", convertedLines);
+			
+			
 		}
 
 		private static IEnumerable<string> PrepareLines(string filename)
@@ -79,23 +107,19 @@ namespace Exceptions
 			yield return lineIndex.ToString();
 		}
 
-	    public static string ConvertLine(string arg) 
-		{												
-			try
-			{
-				return ConvertAsDateTime(arg);
-			}
-			catch
-			{
-				try
-				{
-					return ConvertAsDouble(arg);
-				}
-				catch
-				{
-					return ConvertAsCharIndexInstruction(arg);
-				}
-			}
+	    public static string ConvertLine(string arg)
+	    {
+		    if (TryConvertAsDateTime(arg))
+			    return ConvertAsDateTime(arg);
+		    if (TryConvertAsDouble(arg))
+			    return ConvertAsDouble(arg);
+			return ConvertAsCharIndexInstruction(arg) ?? throw new FormatException("Некорректная строка");
+	    }
+
+		private static bool TryConvertAsCharIndexInstruction(string s)
+		{
+			var parts = s.Split();
+			return int.TryParse(parts[0], out var ans);
 		}
 
 		private static string ConvertAsCharIndexInstruction(string s)
@@ -109,14 +133,26 @@ namespace Exceptions
 			return text[charIndex].ToString();
 		}
 
+		private static bool TryConvertAsDateTime(string arg)
+		{
+			return DateTime.TryParse(arg, out var ans);
+		}
+
 		private static string ConvertAsDateTime(string arg)
 		{
-			return DateTime.Parse(arg).ToString(CultureInfo.InvariantCulture);
+			DateTime.TryParse(arg, out var ans);
+
+			return ans.ToString(CultureInfo.InvariantCulture);
+		}
+		private static bool TryConvertAsDouble(string arg)
+		{
+			return double.TryParse(arg, out var ans);
 		}
 
 		private static string ConvertAsDouble(string arg)
 		{
-			return double.Parse(arg).ToString(CultureInfo.InvariantCulture);
+			double.TryParse(arg, out var ans);
+			return ans.ToString(CultureInfo.InvariantCulture);
 		}
 	}
 }
